+++ src/core/file.hpp
@@ -13,12 +13,12 @@ public:
   File& operator=(File&&);
 
   int Fd() const;
-  int Size() const;
+  size_t Size() const;
   bool Ok() const;
 
 private:
   int fd;
-  int size;
+  size_t size;
 };
 
 }  // namespace os
